+++ lib/honeybadger/notice.rb
@@ -305,7 +305,7 @@ module Honeybadger
       return nil unless url =~ /\S/
 
       url = url.dup
-      url.scan(/&([^=]+)=([^&]+)/).each do |m|
+      url.scan(/(?:^|&|\?)([^=?&]+)=([^&]+)/).each do |m|
         next unless filter_key?(m[0])
         url.gsub!(/#{m[1]}/, '[FILTERED]')
       end
@@ -321,10 +321,18 @@ module Honeybadger
       if cgi_data
         clean_unserializable_data_from(:cgi_data)
         filter(cgi_data)
+        filter_cgi_data_params(cgi_data)
       end
       if session_data
         clean_unserializable_data_from(:session_data)
         filter(session_data)
+      end
+    end
+
+    def filter_cgi_data_params(cgi_data)
+      cgi_data.each_pair do |key, value|
+        next unless value.kind_of?(String) && key.match(/\A[A-Z_]+\Z/)
+        cgi_data[key] = filter_url(value)
       end
     end